+++ src/main/java/insane96mcp/enhancedai/modules/zombie/ai/BlockBreakerGoal.java
@@ -1,6 +1,6 @@
 package insane96mcp.enhancedai.modules.zombie.ai;
 
-import insane96mcp.enhancedai.modules.humanoid.MinerMobs;
+import insane96mcp.enhancedai.modules.mobs.MinerMobs;
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.sounds.SoundSource;